Refactoring, retiring, and renewing some tutorials
